+++ docs/src/examples/CheckboxSlots.vue
@@ -1,19 +1,30 @@
 <template>
-  <nhs-checkboxes
-    :items="items" v-model="model"
-    label="What is your nationality?"
-    hint="If you have more than 1 nationality, select all options that are relevant to you."
-  >
+  <div>
+    <nhs-checkboxes
+      :items="items" v-model="model"
+      label="What is your nationality?"
+      hint="If you have more than 1 nationality, select all options that are relevant to you."
+      ref="checkbox" :rules="rules"
+    >
 
-    <template #hint="hint">
-      <a>{{hint.props}}</a>
-    </template>
+      <template #hint="hint">
+        <a>{{hint.props}}</a>
+      </template>
 
-    <template #item-label="item">
-      <h3>{{item.props.label}}</h3>
-    </template>
+      <template #item-label="item">
+        <h3>{{item.props.label}}</h3>
+      </template>
 
-  </nhs-checkboxes>
+    </nhs-checkboxes>
+    <nhs-row>
+      <nhs-col :span="33">
+        <nhs-button @click="validate()">Validate</nhs-button>
+      </nhs-col>
+      <nhs-col :span="33">
+        <nhs-button @click="reset()">Reset</nhs-button>
+      </nhs-col>
+    </nhs-row>
+  </div>
 </template>
 
 <script>
@@ -34,8 +45,26 @@
             value: 'other'
           }
         ],
-        model: []
+        model: [],
+        rules: [
+          (v) => v.length > 0 || 'Please select an option'
+        ]
       }
+    },
+    methods: {
+      validate() {
+        var valid = this.$refs.checkbox.validate()
+        alert(`Valid value: ${valid}`)
+      },
+      reset() {
+        this.model = [
+          'british'
+        ]
+        this.$refs.checkbox.resetError()
+      }
+    },
+    mounted() {
+      this.$refs.checkbox.validate()
     }
   }
 </script>