+++ docs/src/examples/DateInputSlots.vue
@@ -17,7 +17,14 @@
         {{label}}
       </template>
     </nhs-input-group>
-    <nhs-button @click="validate()">Validate</nhs-button>
+    <nhs-row>
+      <nhs-col :span="33">
+        <nhs-button @click="validate()">Validate</nhs-button>
+      </nhs-col>
+      <nhs-col :span="33">
+        <nhs-button @click="reset()">Reset</nhs-button>
+      </nhs-col>
+    </nhs-row>
   </div>
 </template>
 
@@ -55,6 +62,10 @@
       validate() {
         var valid = this.$refs.input.validate()
         alert(`Valid value: ${valid}`)
+      },
+      reset() {
+        this.model = {}
+        this.$refs.input.resetError()
       }
     }
   }