+++ docs/src/examples/InputSlots.vue
@@ -4,7 +4,7 @@
       ref="input"
       label="National Insurance number"
       hint="It’s on your National Insurance card, benefit letter, payslip or P60. For example, ‘QQ 12 34 56 C’."
-      :rules="rules"
+      :rules="rules" v-model="model"
     >
       <template #label="label">
         <h1>{{label.props}}</h1>
@@ -25,6 +25,9 @@
       <nhs-col :span="33">
         <nhs-button @click="validate()">Validate</nhs-button>
       </nhs-col>
+      <nhs-col :span="33">
+        <nhs-button @click="reset()">Reset</nhs-button>
+      </nhs-col>
     </nhs-row>
   </div>
 </template>
@@ -35,7 +38,8 @@
       return {
         rules: [
           (v) => !!v || 'Error message goes here'
-        ]
+        ],
+        model: ''
       }
     },
     mounted() {
@@ -45,6 +49,10 @@
       validate() {
         var valid = this.$refs.input.validate()
         alert(`Valid value: ${valid}`)
+      },
+      reset() {
+        this.model = ''
+        this.$refs.input.resetError()
       }
     }
   }