+++ docs/src/views/form/Radio.vue
@@ -104,6 +104,10 @@
           {
             name: 'validate',
             description: 'Validates v-model value based on rules. Returns true if valid'
+          },
+          {
+            name: 'resetError',
+            description: 'reset error message'
           }
         ]
       }